requirements.txt
pandas
rdflib
# anndata
neo4j
pyyaml